+++ app/src/main/java/au/com/holberton/simplicity/mobile/productlisting/ListingSummary.kt
@@ -10,7 +10,7 @@ data class ListingSummary(
     val name: String,
     val category: String,
     val quantity: Int?,
-    val upc: Double,
+    val upc: Long,
     val costPrice: Double,
     val salePrice: Double,
     val location: String,
